Added visible_changes method
Visible changes can be used to select only changes where the change is significant, i.e. a change from “3” to 3 is not visible because the string representation is the same. Similarly, a change from {} to nil is not significant because both fail the `present?` check. A change from {} to “3” would be visible because the string is not empty.