Support associations with custom primary key
If an association uses a non-standard primary key, we won't be able to find it by the changed value.

Instead, we use the primary key config stored in the reflection we have for this association to determine which column to use as the primary key to look up the new associated value.